add additional tests
